models/image/cnndetection/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
torch==2.0.1
torchvision==0.15.2
pillow==10.0.1
numpy
scipy
scikit-learn
opencv-python
python-multipart